vmcleaner: move tests around
